# swagger.yaml
openapi: 3.0.3
info:
  title: Crypto Trading Platform API
  description: |
    Complete API documentation for the crypto trading platform.
    Includes REST APIs and WebSocket connections for real-time data.
    
    Base URL: http://192.168.1.47:8080
    WebSocket URL: ws://192.168.1.47:8081/ws
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://192.168.1.47:8080
    description: Backend API Server
  - url: ws://192.168.1.47:8081
    description: WebSocket Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or email verification
  schemas:
...
    BuyTradeRequest:
      type: object
      required:
        - chain_id
        - token_address
        - amount
        - slippage_bps
      properties:
        chain_id:
          type: integer
          description: Blockchain chain ID (501 for Solana)
          example: 501
        token_address:
          type: string
          description: Token contract address
          example: "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
        amount:
          type: string
          description: Amount to buy in native currency (SOL)
          example: "0.1"
        slippage_bps:
          type: integer
          description: Slippage tolerance in basis points (0-10000)
          minimum: 0
          maximum: 10000
          example: 500
    BuyTradeResponse:
      type: object
      properties:
        success:
          type: boolean
        trade_id:
          type: string
          description: Unique trade identifier
        message:
          type: string
        error:
          type: string
    SellTradeRequest:
      type: object
      required:
        - chain_id
        - token_address
        - amount
      properties:
        chain_id:
          type: integer
          description: Blockchain chain ID (501 for Solana)
          example: 501
        token_address:
          type: string
          description: Token contract address
        amount:
          type: string
          description: Amount of tokens to sell
          example: "100.5"
    SellTradeResponse:
      type: object
      properties:
        success:
          type: boolean
        trade_id:
          type: string
        message:
          type: string
        error:
          type: string
          
    # Wallet Schemas
    WalletBalance:
      type: object
      properties:
        chain_id:
          type: integer
        balance:
          type: string
          description: Balance in native currency
        currency:
          type: string
          example: "SOL"
    WalletBalanceResponse:
      type: object
      properties:
        success:
          type: boolean
        balances:
          type: array
          items:
            $ref: '#/components/schemas/WalletBalance'
        message:
          type: string
        status:
          type: integer
    DepositResponse:
      type: object
      properties:
        status:
          type: integer
        data:
          type: object
          properties:
            wallet_address:
              type: string
              description: Deposit wallet address
            chain_id:
              type: integer
        message:
          type: string
    WithdrawRequest:
      type: object
      required:
        - chain_id
        - to_wallet
        - amount
      properties:
        chain_id:
          type: integer
          example: 501
        to_wallet:
          type: string
          description: Destination wallet address
        amount:
          type: string
          description: Amount to withdraw
    WithdrawResponse:
      type: object
      properties:
        status:
          type: integer
        data:
          type: object
          properties:
            success:
              type: boolean
            tx_hash:
              type: string
              description: Transaction hash
        message:
          type: string

    # Token Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
paths:
...
  /api/trade/buy:
    post:
      tags:
        - Trading
      summary: Buy tokens
      description: Execute a buy order for specified token
      operationId: buyToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyTradeRequest'
      responses:
        '200':
          description: Buy order executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyTradeResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/trade/sell:
    post:
      tags:
        - Trading
      summary: Sell tokens
      description: Execute a sell order for specified token
      operationId: sellToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellTradeRequest'
      responses:
        '200':
          description: Sell order executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellTradeResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Wallet Endpoints
  /api/wallet/balance:
    get:
      tags:
        - Wallet
      summary: Get wallet balance
      description: Retrieve user's wallet balances for all chains
      operationId: getWalletBalance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalanceResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/deposit:
    get:
      tags:
        - Wallet
      summary: Get deposit address
      description: Get wallet address for depositing funds
      operationId: getDepositAddress
      security:
        - BearerAuth: []
      parameters:
        - name: chain_id
          in: query
          required: true
          schema:
            type: integer
            example: 501
          description: Blockchain chain ID
      responses:
        '200':
          description: Deposit address retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
        '400':
          description: Missing chain_id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/withdraw:
    post:
      tags:
        - Wallet
      summary: Withdraw funds
      description: Withdraw funds from wallet to external address
      operationId: withdrawFunds
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        '200':
          description: Withdrawal successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
  # Token Endpoints
...